src/reservationservice/pyproject.toml
[tool.poetry]
name = "reservation-service"
version = "0.1.0"
description = ""
authors = ["Daniil"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.111.0"}
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
aiohttp = "^3.10.5"
pyjwt = "^2.9.0"

[tool.poetry.group.test.dependencies]
pytest = "8.2.0"
pytest-asyncio = "0.21.2"
pylint = "^3.2.5"
aioresponses = "^0.7.6"
pytest-env = "1.1.3"
pytest-cov = "^5.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
